pyproject.toml
[project]
name = "model_choice"
version = "1.4.0"
description = "Universal LLM model selector and caller -- pick the right model for the job"
requires-python = ">=3.10"
dependencies = ["litellm", "pyyaml"]

[project.scripts]
model_choice = "model_choice.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["model_choice*"]
